requirements/test.txt
# test (integ) environment requirements

# includes the base.txt requirements needed in all environments
-r base.txt   
 
# below are requirements specific to the test/integ environment

django-debug-toolbar==0.11.0
git+https://github.com/inducer/pudb.git#egg=pudb
gunicorn==0.16.1

newrelic==1.12.0.56

git+ssh://git@git.huit.harvard.edu/icommons-python/django-icommons-common.git@release_0.8#egg=django-icommons-common
